DRY test files
